+++ dio/CHANGELOG.md
@@ -7,6 +7,7 @@ See the [Migration Guide][] for the complete breaking changes list.**
 
 - Fix `receiveTimeout` throws exception after the request has been cancelled.
 - Catch sync/async exceptions in interceptors' handlers.
+- Throws precise `StateError` for handler's duplicated calls.
 
 ## 5.4.1